+++ WarriorsSnuggery/Map/Piece.cs
@@ -85,7 +85,7 @@ namespace WarriorsSnuggery.Maps
 								{
 									var split = actor.Key.Split(';');
 									var type = split[0];
-									var team = split.Length > 1 ? int.Parse(split[1]) : 0;
+									var team = split.Length > 1 ? byte.Parse(split[1]) : (byte) 0;
 									var bot = split.Length > 1 ? split[2].Equals("true") : false;
 									actorList.Add(new ActorNode(0, actor.ToCPos(), type, team, 1f, bot));
 								}
